Move `MuiEvent` to `x-internals` package
