+++ packages/x-data-grid/src/models/api/gridEditingApi.ts
@@ -1,9 +1,9 @@
+import { MuiBaseEvent } from '@mui/x-internals/types';
 import { GridCellMode, GridRowMode } from '../gridCell';
 import { GridCellModes, GridRowModes } from '../gridEditRowModel';
 import { GridRowId, GridRowModel } from '../gridRows';
 import { GridCellParams } from '../params/gridCellParams';
 import { GridEditCellValueParams } from '../params/gridEditCellParams';
-import { MuiBaseEvent } from '../muiEvent';
 
 export type GridCellModesModelProps =
   | ({ mode: GridCellModes.View } & Omit<GridStopCellEditModeParams, 'id' | 'field'>)
