+++ packages/x-data-grid/src/models/events/gridEventPublisher.ts
@@ -1,4 +1,4 @@
-import type { MuiBaseEvent } from '../muiEvent';
+import type { MuiBaseEvent } from '@mui/x-internals/types';
 import type { GridEventLookup, GridEvents } from './gridEventLookup';
 
 type PublisherArgsNoEvent<E extends GridEvents, T extends { params: any }> = [E, T['params']];
