+++ packages/x-internals/src/types/index.ts
@@ -2,6 +2,7 @@ export * from './AppendKeys';
 export * from './DefaultizedProps';
 export * from './MakeOptional';
 export * from './MakeRequired';
+export * from './MuiEvent';
 export * from './PrependKeys';
 export * from './RefObject';
 export * from './SlotComponentPropsFromProps';
