+++ packages/x-tree-view/src/internals/models/events.ts
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import { MuiEvent } from '@mui/x-internals/types';
 
 export interface TreeViewEventLookupElement {
   params: object;
@@ -8,12 +8,3 @@ export type TreeViewEventListener<E extends TreeViewEventLookupElement> = (
   params: E['params'],
   event: MuiEvent<{}>,
 ) => void;
-
-export type MuiBaseEvent =
-  | React.SyntheticEvent<HTMLElement>
-  | DocumentEventMap[keyof DocumentEventMap]
-  | {};
-
-export type MuiEvent<E extends MuiBaseEvent = MuiBaseEvent> = E & {
-  defaultMuiPrevented?: boolean;
-};
